@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.1.4 <=4.0.0"
 
 dependencies:
   flutter:
@@ -34,10 +34,11 @@ dependencies:
   equatable: ^2.0.2
   google_fonts: ^5.1.0
   flutter_rating_bar: ^4.0.0
-  cached_network_image: ^3.0.0
+  cached_network_image: ^3.3.0
   get_it: ^7.1.3
   sqflite: ^2.0.0+3
   path_provider: ^2.0.2
+  intl: ^0.17.0
 
 dev_dependencies:
   mockito: ^5.0.8
