@@ -1,26 +1,26 @@
 import 'package:dartz/dartz.dart';
 import 'package:ditonton/common/failure.dart';
 import 'package:ditonton/common/state_enum.dart';
-import 'package:ditonton/domain/usecases/get_watchlist_movies.dart';
-import 'package:ditonton/presentation/provider/watchlist_movie_notifier.dart';
+import 'package:ditonton/domain/usecases/watchlist/get_watchlist.dart';
+import 'package:ditonton/presentation/provider/watchlist_notifier.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../dummy_data/dummy_objects.dart';
-import 'watchlist_movie_notifier_test.mocks.dart';
+import 'watchlist_notifier_test.mocks.dart';
 
-@GenerateMocks([GetWatchlistMovies])
+@GenerateMocks([GetWatchlist])
 void main() {
-  late WatchlistMovieNotifier provider;
-  late MockGetWatchlistMovies mockGetWatchlistMovies;
+  late WatchlistNotifier provider;
+  late MockGetWatchlist mockGetWatchlist;
   late int listenerCallCount;
 
   setUp(() {
     listenerCallCount = 0;
-    mockGetWatchlistMovies = MockGetWatchlistMovies();
-    provider = WatchlistMovieNotifier(
-      getWatchlistMovies: mockGetWatchlistMovies,
+    mockGetWatchlist = MockGetWatchlist();
+    provider = WatchlistNotifier(
+      getWatchlist: mockGetWatchlist,
     )..addListener(() {
         listenerCallCount += 1;
       });
@@ -28,19 +28,19 @@ void main() {
 
   test('should change movies data when data is gotten successfully', () async {
     // arrange
-    when(mockGetWatchlistMovies.execute())
-        .thenAnswer((_) async => Right([testWatchlistMovie]));
+    when(mockGetWatchlist.execute())
+        .thenAnswer((_) async => Right([testWatchlist]));
     // act
     await provider.fetchWatchlistMovies();
     // assert
     expect(provider.watchlistState, RequestState.Loaded);
-    expect(provider.watchlistMovies, [testWatchlistMovie]);
+    expect(provider.watchlistData, [testWatchlist]);
     expect(listenerCallCount, 2);
   });
 
   test('should return error when data is unsuccessful', () async {
     // arrange
-    when(mockGetWatchlistMovies.execute())
+    when(mockGetWatchlist.execute())
         .thenAnswer((_) async => Left(DatabaseFailure("Can't get data")));
     // act
     await provider.fetchWatchlistMovies();
