@@ -1,12 +1,11 @@
 import 'package:ditonton/common/exception.dart';
 import 'package:ditonton/data/datasources/db/database_helper.dart';
-import 'package:ditonton/data/models/movie_table.dart';
+import 'package:ditonton/data/models/watchlist_table.dart';
 
 abstract class MovieLocalDataSource {
-  Future<String> insertWatchlist(MovieTable movie);
-  Future<String> removeWatchlist(MovieTable movie);
-  Future<MovieTable?> getMovieById(int id);
-  Future<List<MovieTable>> getWatchlistMovies();
+  Future<String> insertWatchlist(WatchlistTable movie);
+  Future<String> removeWatchlist(WatchlistTable movie);
+  Future<WatchlistTable?> getWatchlistById(int id);
 }
 
 class MovieLocalDataSourceImpl implements MovieLocalDataSource {
@@ -15,7 +14,7 @@ class MovieLocalDataSourceImpl implements MovieLocalDataSource {
   MovieLocalDataSourceImpl({required this.databaseHelper});
 
   @override
-  Future<String> insertWatchlist(MovieTable movie) async {
+  Future<String> insertWatchlist(WatchlistTable movie) async {
     try {
       await databaseHelper.insertWatchlist(movie);
       return 'Added to Watchlist';
@@ -25,7 +24,7 @@ class MovieLocalDataSourceImpl implements MovieLocalDataSource {
   }
 
   @override
-  Future<String> removeWatchlist(MovieTable movie) async {
+  Future<String> removeWatchlist(WatchlistTable movie) async {
     try {
       await databaseHelper.removeWatchlist(movie);
       return 'Removed from Watchlist';
@@ -35,18 +34,12 @@ class MovieLocalDataSourceImpl implements MovieLocalDataSource {
   }
 
   @override
-  Future<MovieTable?> getMovieById(int id) async {
-    final result = await databaseHelper.getMovieById(id);
+  Future<WatchlistTable?> getWatchlistById(int id) async {
+    final result = await databaseHelper.getWatchlistById(id);
     if (result != null) {
-      return MovieTable.fromMap(result);
+      return WatchlistTable.fromMap(result);
     } else {
       return null;
     }
-  }
-
-  @override
-  Future<List<MovieTable>> getWatchlistMovies() async {
-    final result = await databaseHelper.getWatchlistMovies();
-    return result.map((data) => MovieTable.fromMap(data)).toList();
   }
 }
