@@ -1,24 +1,45 @@
 import 'package:ditonton/data/datasources/db/database_helper.dart';
 import 'package:ditonton/data/datasources/movie_local_data_source.dart';
 import 'package:ditonton/data/datasources/movie_remote_data_source.dart';
+import 'package:ditonton/data/datasources/tv_series_local_data_source.dart';
+import 'package:ditonton/data/datasources/tv_series_remote_data_source.dart';
+import 'package:ditonton/data/datasources/watchlist_local_data_source.dart';
 import 'package:ditonton/data/repositories/movie_repository_impl.dart';
+import 'package:ditonton/data/repositories/tv_series_repository_impl.dart';
+import 'package:ditonton/data/repositories/watchlist_repository_impl.dart';
 import 'package:ditonton/domain/repositories/movie_repository.dart';
+import 'package:ditonton/domain/repositories/tv_series_repository.dart';
+import 'package:ditonton/domain/repositories/watchlist_repository.dart';
 import 'package:ditonton/domain/usecases/get_movie_detail.dart';
 import 'package:ditonton/domain/usecases/get_movie_recommendations.dart';
 import 'package:ditonton/domain/usecases/get_now_playing_movies.dart';
 import 'package:ditonton/domain/usecases/get_popular_movies.dart';
 import 'package:ditonton/domain/usecases/get_top_rated_movies.dart';
-import 'package:ditonton/domain/usecases/get_watchlist_movies.dart';
 import 'package:ditonton/domain/usecases/get_watchlist_status.dart';
 import 'package:ditonton/domain/usecases/remove_watchlist.dart';
 import 'package:ditonton/domain/usecases/save_watchlist.dart';
 import 'package:ditonton/domain/usecases/search_movies.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_airing_today_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_on_the_air_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_popular_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_top_rated_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_tv_series_detail.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_tv_series_recommendation.dart';
+import 'package:ditonton/domain/usecases/tv_series/get_watchlist_status_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/remove_watchlist_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/save_watchlist_tv_series.dart';
+import 'package:ditonton/domain/usecases/tv_series/search_tv_series.dart';
+import 'package:ditonton/domain/usecases/watchlist/get_watchlist.dart';
 import 'package:ditonton/presentation/provider/movie_detail_notifier.dart';
 import 'package:ditonton/presentation/provider/movie_list_notifier.dart';
 import 'package:ditonton/presentation/provider/movie_search_notifier.dart';
 import 'package:ditonton/presentation/provider/popular_movies_notifier.dart';
+import 'package:ditonton/presentation/provider/search_tv_series_notifier.dart';
 import 'package:ditonton/presentation/provider/top_rated_movies_notifier.dart';
-import 'package:ditonton/presentation/provider/watchlist_movie_notifier.dart';
+import 'package:ditonton/presentation/provider/tv_series_detail_notifier.dart';
+import 'package:ditonton/presentation/provider/tv_series_list_notifier.dart';
+import 'package:ditonton/presentation/provider/tv_series_more_notifier.dart';
+import 'package:ditonton/presentation/provider/watchlist_notifier.dart';
 import 'package:http/http.dart' as http;
 import 'package:get_it/get_it.dart';
 
@@ -58,9 +79,39 @@ void init() {
     ),
   );
   locator.registerFactory(
-    () => WatchlistMovieNotifier(
-      getWatchlistMovies: locator(),
+    () => WatchlistNotifier(
+      getWatchlist: locator(),
     ),
+  );
+  locator.registerFactory(
+    () => TVSeriesListNotifier(
+      getAirTodayTVSeries: locator(),
+      getTopRatedTVSeries: locator(),
+      getOnTheAirTVSeries: locator(),
+      getPopularTVSeries: locator(),
+    ),
+  );
+
+  locator.registerFactory(
+    () => TVSeriesDetailNotifier(
+      getTVSeriesDetail: locator(),
+      getTVSeriesRecommendations: locator(),
+      saveWatchlist: locator(),
+      removeWatchlist: locator(),
+      getWatchListStatus: locator(),
+    ),
+  );
+
+  locator.registerFactory(
+    () => TVSeriesMoreNotifier(
+      getTopRatedTVSeries: locator(),
+      getAirTodayTVSeries: locator(),
+      getPopularTVSeries: locator(),
+    ),
+  );
+
+  locator.registerFactory(
+    () => SearchTVSeriesNotifier(searchTVSeries: locator())
   );
 
   // use case
@@ -73,7 +124,17 @@ void init() {
   locator.registerLazySingleton(() => GetWatchListStatus(locator()));
   locator.registerLazySingleton(() => SaveWatchlist(locator()));
   locator.registerLazySingleton(() => RemoveWatchlist(locator()));
-  locator.registerLazySingleton(() => GetWatchlistMovies(locator()));
+  locator.registerLazySingleton(() => GetAirTodayTVSeries(locator()));
+  locator.registerLazySingleton(() => GetTopRatedTVSeries(locator()));
+  locator.registerLazySingleton(() => GetOnTheAirTVSeries(locator()));
+  locator.registerLazySingleton(() => GetPopularTVSeries(locator()));
+  locator.registerLazySingleton(() => GetTVSeriesDetail(locator()));
+  locator.registerLazySingleton(() => GetTVSeriesRecommendations(locator()));
+  locator.registerLazySingleton(() => GetWatchlist(locator()));
+  locator.registerLazySingleton(() => RemoveWatchlistTVSeries(locator()));
+  locator.registerLazySingleton(() => SaveWatchlistTVSeries(locator()));
+  locator.registerLazySingleton(() => GetWatchListStatusTVSeries(locator()));
+  locator.registerLazySingleton(() => SearchTVSeries(locator()));
 
   // repository
   locator.registerLazySingleton<MovieRepository>(
@@ -82,12 +143,27 @@ void init() {
       localDataSource: locator(),
     ),
   );
+  locator.registerLazySingleton<TVSeriesRepository>(
+    () => TVSeriesRepositoryImpl(
+      remoteDataSource: locator(),
+      localDataSource: locator(),
+    ),
+  );
+  locator.registerLazySingleton<WatchlistRepository>(
+    () => WatchlistRepositoryImpl(localDataSource: locator())
+  );
 
   // data sources
   locator.registerLazySingleton<MovieRemoteDataSource>(
       () => MovieRemoteDataSourceImpl(client: locator()));
   locator.registerLazySingleton<MovieLocalDataSource>(
       () => MovieLocalDataSourceImpl(databaseHelper: locator()));
+  locator.registerLazySingleton<WatchlistLocalDataSource>(
+      () => WatchlistLocalDataSourceImpl(databaseHelper: locator()));
+  locator.registerLazySingleton<TVSeriesRemoteDataSource>(
+      () => TVSeriesRemoteDataSourceImpl(client: locator()));
+  locator.registerLazySingleton<TVSeriesLocalDataSource>(
+      () => TVSeriesLocalDataSourceImpl(databaseHelper: locator()));
 
   // helper
   locator.registerLazySingleton<DatabaseHelper>(() => DatabaseHelper());
