@@ -1,9 +1,9 @@
-// Mocks generated by Mockito 5.0.8 from annotations
+// Mocks generated by Mockito 5.4.2 from annotations
 // in ditonton/test/presentation/pages/popular_movies_page_test.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i6;
-import 'dart:ui' as _i7;
 
 import 'package:ditonton/common/state_enum.dart' as _i4;
 import 'package:ditonton/domain/entities/movie.dart' as _i5;
@@ -12,13 +12,27 @@ import 'package:ditonton/presentation/provider/popular_movies_notifier.dart'
     as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
-class _FakeGetPopularMovies extends _i1.Fake implements _i2.GetPopularMovies {}
+class _FakeGetPopularMovies_0 extends _i1.SmartFake
+    implements _i2.GetPopularMovies {
+  _FakeGetPopularMovies_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
 /// A class which mocks [PopularMoviesNotifier].
 ///
@@ -30,41 +44,72 @@ class MockPopularMoviesNotifier extends _i1.Mock
   }
 
   @override
-  _i2.GetPopularMovies get getPopularMovies =>
-      (super.noSuchMethod(Invocation.getter(#getPopularMovies),
-          returnValue: _FakeGetPopularMovies()) as _i2.GetPopularMovies);
+  _i2.GetPopularMovies get getPopularMovies => (super.noSuchMethod(
+        Invocation.getter(#getPopularMovies),
+        returnValue: _FakeGetPopularMovies_0(
+          this,
+          Invocation.getter(#getPopularMovies),
+        ),
+      ) as _i2.GetPopularMovies);
   @override
-  _i4.RequestState get state => (super.noSuchMethod(Invocation.getter(#state),
-      returnValue: _i4.RequestState.Empty) as _i4.RequestState);
+  _i4.RequestState get state => (super.noSuchMethod(
+        Invocation.getter(#state),
+        returnValue: _i4.RequestState.Empty,
+      ) as _i4.RequestState);
   @override
-  List<_i5.Movie> get movies => (super.noSuchMethod(Invocation.getter(#movies),
-      returnValue: <_i5.Movie>[]) as List<_i5.Movie>);
+  List<_i5.Movie> get movies => (super.noSuchMethod(
+        Invocation.getter(#movies),
+        returnValue: <_i5.Movie>[],
+      ) as List<_i5.Movie>);
   @override
-  String get message =>
-      (super.noSuchMethod(Invocation.getter(#message), returnValue: '')
-          as String);
+  String get message => (super.noSuchMethod(
+        Invocation.getter(#message),
+        returnValue: '',
+      ) as String);
   @override
-  bool get hasListeners =>
-      (super.noSuchMethod(Invocation.getter(#hasListeners), returnValue: false)
-          as bool);
+  bool get hasListeners => (super.noSuchMethod(
+        Invocation.getter(#hasListeners),
+        returnValue: false,
+      ) as bool);
   @override
-  _i6.Future<void> fetchPopularMovies() =>
-      (super.noSuchMethod(Invocation.method(#fetchPopularMovies, []),
-          returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future.value()) as _i6.Future<void>);
+  _i6.Future<void> fetchPopularMovies() => (super.noSuchMethod(
+        Invocation.method(
+          #fetchPopularMovies,
+          [],
+        ),
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
   @override
-  void addListener(_i7.VoidCallback? listener) =>
-      super.noSuchMethod(Invocation.method(#addListener, [listener]),
-          returnValueForMissingStub: null);
+  void addListener(dynamic listener) => super.noSuchMethod(
+        Invocation.method(
+          #addListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  void removeListener(_i7.VoidCallback? listener) =>
-      super.noSuchMethod(Invocation.method(#removeListener, [listener]),
-          returnValueForMissingStub: null);
+  void removeListener(dynamic listener) => super.noSuchMethod(
+        Invocation.method(
+          #removeListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  void dispose() => super.noSuchMethod(Invocation.method(#dispose, []),
-      returnValueForMissingStub: null);
+  void dispose() => super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  void notifyListeners() =>
-      super.noSuchMethod(Invocation.method(#notifyListeners, []),
-          returnValueForMissingStub: null);
+  void notifyListeners() => super.noSuchMethod(
+        Invocation.method(
+          #notifyListeners,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
 }
